separated filter implementation from testing
