+++ capacitiveSensor/arduino/testFilter/movingFilter.cpp
@@ -1,89 +1,9 @@
 #include <stdlib.h>
 #include <stdio.h>
-#include <cmath>
 #include <iostream>
 #include <fstream>
 #include <string>
-#include "movingFilter.h"
-
-// Filter constructor
-MovingFilter::MovingFilter(int n, double thresh, double alpha) {
-    // Set parameters
-    this->n = n;
-    this->thresh = thresh;
-    this->alpha = alpha;
-
-    // Set variables
-    t = 0;
-    p = 0;
-    mean = 0.0;
-    std = 0.0;
-
-    // Allocate buffer
-    buffer = new double[n];
-}
-
-// Filter methods
-int MovingFilter::applyFilter(double y) { 
-    // Signal value
-    int x;
-    
-    // Cache old value
-    double bufferOld = buffer[p];
-    double meanOld = mean;
-
-    // Check for signal (if buffer filled)
-    if ((t > n) && (abs(y - mean) > (thresh*std))) {
-        // Update signal array
-        if ((y - mean) > 0.0) { // positve signal
-            x = 1;
-        }
-        else { // negative signal
-            x = -1;
-        }
-
-        // Update buffer
-        int prev = ((p - 1) + n) % n; // wraps to end if needed
-        buffer[p] = (alpha*y) + ((1.0 - alpha)*buffer[prev]);
-    }
-    else {
-        // Update signal and buffer
-        x = 0;
-        buffer[p] = y;
-    }
-
-    // Update buffer stats
-    if (t % 100 == 0) {
-        // Get ground truth stats occasionally to correct drift
-        mean = 0;
-        for (int i = 0; i < n; i++) {
-            mean += buffer[i];
-        }
-        mean = mean / (double) n;
-        std = 0;
-        for (int i = 0; i < n; i++){
-            std += pow((buffer[i] - mean), 2);
-        }
-        std = sqrt(std / (double) n);
-    }
-    else {
-        // Use single point updates for faster processing
-        double dMean = (1.0/(double) n)*(buffer[p] - bufferOld);
-        mean += dMean;
-        std = (pow(std, 2) - (1.0/(double) n)*pow(bufferOld - meanOld, 2) 
-               + (1.0/(double) n)*pow(buffer[p] - mean, 2) 
-               + (2*dMean/(double) n)*(bufferOld - meanOld) 
-               + (((double) n - 1)/((double) n))*pow(dMean, 2));
-        std = sqrt(std);
-    }
-
-    // Increment indices
-    p = (p + 1) % n;
-    t++;
-
-    // Return signal
-    return x;
-}
+#include "filter.h"
 
 int main(int argc, char *argv[]) {
     // Create new filter
